nimble/audio/bass: Fix Modify Source operation PDU parsing
Num_Subgroups is 1 byte so adjust accordingly. Also validate received
data while parising it. In case of error reject operation.

+++ nimble/host/audio/services/bass/include/services/bass/ble_audio_svc_bass.h
@@ -325,7 +325,7 @@ struct ble_svc_audio_bass_operation {
             uint16_t pa_interval;
 
             /** Number of subgroups */
-            uint16_t num_subgroups;
+            uint8_t num_subgroups;
 
             /** Subgroup entries */
             struct ble_svc_audio_bass_subgroup
